Moved copy logic from ProblemsBusinessService to ProblemGroupsBusinessService https://github.com/SoftUni-Internal/suls-issues/issues/5006
